# generated/csharp/src/Org.OpenAPITools/Controllers/TenantUsersApi.cs
/*
 * (title)
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Org.OpenAPITools.Attributes;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TenantUsersApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="userId"></param>
        /// <param name="tenantUsersAssignRolesRequest"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpPut]
        [Route("/api/v1/tenants/{tenantId}/users/{userId}/roles")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("TenantUsersAssignRoles")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantUsersAssignRoles([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "userId")][Required]string userId, [FromBody]TenantUsersAssignRolesRequest tenantUsersAssignRolesRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"username\" : \"username\",\r\n  \"email\" : \"email\",\r\n  \"displayName\" : \"displayName\",\r\n  \"status\" : \"active\",\r\n  \"roleIds\" : [ \"roleIds\", \"roleIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="userId"></param>
        /// <param name="tenantUsersChangeUserStatusRequest"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpPatch]
        [Route("/api/v1/tenants/{tenantId}/users/{userId}/status")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("TenantUsersChangeUserStatus")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantUsersChangeUserStatus([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "userId")][Required]string userId, [FromBody]TenantUsersChangeUserStatusRequest tenantUsersChangeUserStatusRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"username\" : \"username\",\r\n  \"email\" : \"email\",\r\n  \"displayName\" : \"displayName\",\r\n  \"status\" : \"active\",\r\n  \"roleIds\" : [ \"roleIds\", \"roleIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="createUserRequest"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpPost]
        [Route("/api/v1/tenants/{tenantId}/users")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("TenantUsersCreateUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantUsersCreateUser([FromRoute (Name = "tenantId")][Required]string tenantId, [FromBody]CreateUserRequest createUserRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"username\" : \"username\",\r\n  \"email\" : \"email\",\r\n  \"displayName\" : \"displayName\",\r\n  \"status\" : \"active\",\r\n  \"roleIds\" : [ \"roleIds\", \"roleIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="userId"></param>
        /// <response code="204">There is no content to send for this request, but the headers may be useful. </response>
        /// <response code="0">An unexpected error response.</response>
        [HttpDelete]
        [Route("/api/v1/tenants/{tenantId}/users/{userId}")]
        [ValidateModelState]
        [SwaggerOperation("TenantUsersDeleteUser")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantUsersDeleteUser([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "userId")][Required]string userId)
        {

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);

            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="userId"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpGet]
        [Route("/api/v1/tenants/{tenantId}/users/{userId}")]
        [ValidateModelState]
        [SwaggerOperation("TenantUsersGetUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantUsersGetUser([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "userId")][Required]string userId)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"username\" : \"username\",\r\n  \"email\" : \"email\",\r\n  \"displayName\" : \"displayName\",\r\n  \"status\" : \"active\",\r\n  \"roleIds\" : [ \"roleIds\", \"roleIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="tenantUsersInviteUserRequest"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpPost]
        [Route("/api/v1/tenants/{tenantId}/users/invitations")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("TenantUsersInviteUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantUsersInviteUser([FromRoute (Name = "tenantId")][Required]string tenantId, [FromBody]TenantUsersInviteUserRequest tenantUsersInviteUserRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"username\" : \"username\",\r\n  \"email\" : \"email\",\r\n  \"displayName\" : \"displayName\",\r\n  \"status\" : \"active\",\r\n  \"roleIds\" : [ \"roleIds\", \"roleIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="page"></param>
        /// <param name="pageSize"></param>
        /// <param name="status"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpGet]
        [Route("/api/v1/tenants/{tenantId}/users")]
        [ValidateModelState]
        [SwaggerOperation("TenantUsersListUsers")]
        [SwaggerResponse(statusCode: 200, type: typeof(TenantUsersListUsers200Response), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantUsersListUsers([FromRoute (Name = "tenantId")][Required]string tenantId, [FromQuery (Name = "page")]int? page, [FromQuery (Name = "pageSize")]int? pageSize, [FromQuery (Name = "status")]UserStatus? status)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"items\" : [ {\r\n    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n    \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n    \"username\" : \"username\",\r\n    \"email\" : \"email\",\r\n    \"displayName\" : \"displayName\",\r\n    \"status\" : \"active\",\r\n    \"roleIds\" : [ \"roleIds\", \"roleIds\" ],\r\n    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n    \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n  }, {\r\n    \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n    \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n    \"username\" : \"username\",\r\n    \"email\" : \"email\",\r\n    \"displayName\" : \"displayName\",\r\n    \"status\" : \"active\",\r\n    \"roleIds\" : [ \"roleIds\", \"roleIds\" ],\r\n    \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n    \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n  } ],\r\n  \"page\" : 0,\r\n  \"pageSize\" : 20,\r\n  \"total\" : 0\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<TenantUsersListUsers200Response>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="tenantId"></param>
        /// <param name="userId"></param>
        /// <param name="updateUserRequest"></param>
        /// <response code="200">The request has succeeded.</response>
        /// <response code="0">An unexpected error response.</response>
        [HttpPatch]
        [Route("/api/v1/tenants/{tenantId}/users/{userId}")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("TenantUsersUpdateUser")]
        [SwaggerResponse(statusCode: 200, type: typeof(User), description: "The request has succeeded.")]
        [SwaggerResponse(statusCode: 0, type: typeof(ErrorResponse), description: "An unexpected error response.")]
        public virtual IActionResult TenantUsersUpdateUser([FromRoute (Name = "tenantId")][Required]string tenantId, [FromRoute (Name = "userId")][Required]string userId, [FromBody]UpdateUserRequest updateUserRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"tenantId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\",\r\n  \"username\" : \"username\",\r\n  \"email\" : \"email\",\r\n  \"displayName\" : \"displayName\",\r\n  \"status\" : \"active\",\r\n  \"roleIds\" : [ \"roleIds\", \"roleIds\" ],\r\n  \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\",\r\n  \"updatedAt\" : \"2000-01-23T04:56:07.000+00:00\"\r\n}";
            exampleJson = "{\r\n  \"code\" : \"BAD_REQUEST\",\r\n  \"message\" : \"message\",\r\n  \"details\" : {\r\n    \"key\" : \"\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<User>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}