# generated/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  version: 0.0.0
...
paths:
...
  /api/v1/tenants/{tenantId}/users:
    get:
      operationId: TenantUsers_listUsers
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserStatus'
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - page
                  - pageSize
                  - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  page:
                    type: integer
                    format: int32
                    example: 0
                  pageSize:
                    type: integer
                    format: int32
                    example: 20
                  total:
                    type: integer
                    format: int64
                    example: 0
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-users
    post:
      operationId: TenantUsers_createUser
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
  /api/v1/tenants/{tenantId}/users/invitations:
    post:
      operationId: TenantUsers_inviteUser
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                roleIds:
                  type: array
                  items:
                    type: string
              required:
                - email
  /api/v1/tenants/{tenantId}/users/{userId}:
    get:
      operationId: TenantUsers_getUser
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-users
    patch:
      operationId: TenantUsers_updateUser
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
    delete:
      operationId: TenantUsers_deleteUser
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-users
  /api/v1/tenants/{tenantId}/users/{userId}/roles:
    put:
      operationId: TenantUsers_assignRoles
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
              required:
                - roleIds
  /api/v1/tenants/{tenantId}/users/{userId}/status:
    patch:
      operationId: TenantUsers_changeUserStatus
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - tenant-users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: '#/components/schemas/UserStatus'
              required:
                - status
components:
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
        email:
          type: string
          format: email
        displayName:
          type: string
        password:
          type: string
        roleIds:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
...
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
        email:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        roleIds:
          type: array
          items:
            type: string
    User:
      type: object
      required:
        - id
        - tenantId
        - username
        - email
        - status
        - roleIds
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        username:
          type: string
          minLength: 1
          maxLength: 64
        email:
          type: string
          format: email
        displayName:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        roleIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserStatus:
      type: string
      enum:
        - active
        - invited
        - suspended
        - disabled
servers:
  - url: https://api.example.com
    description: Production
    variables: {}